remoteproc: ingenic: Lock spinlock before kicking VPU
Add support for the hardware spinlock shared between the VPU and the
main CPU.

In order to implement proper locking, the VPU needs to make sure that no
hardware interrupt will ever happen in a critical section. Since the
only interrupt it can receive (kick from main CPU) is non-maskable, it
needs to make sure that the main CPU won't kick during the critical
section. The hardware spinlock is then used by the VPU firmware to
prevent the main CPU from kicking during a critical section.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>